Experimental code to pull data off a MySQL server
In principle, I think this is all I need; have a key trigger to pull
all the data off the MySQL server and command all the heart objects
accordingly. That key trigger then resets the array to sync with the
database.

That’s the extent of disaster recovery required for the simulation, I
think?