Batch processing for QR code byte arrays, including scale doubling
See QRmaker/README.md for instructions. This isn’t completely
automated, but since it’s theoretically a ‘do once’ sort of job it’s
probably worth preserving flexibility.

Of note: now pixel-doubled, the ‘all.h’ file containing the byte arrays
is >5Mb. This shouldn’t be a problem for Pi/Python, but it’s way too
large for an Arduino… and it’s clearly a C-format header file. So
there’s probably a round of tweaking of the Processing code required
before final output.